refactor: move commands in their own files
